backend/requirements.txt
# Core Django
Django>=4.2,<5
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0

# Environment & Configuration
django-environ>=0.11
django-cors-headers>=4.3

# Authentication
djangorestframework-simplejwt>=5.3

# Media & Files
Pillow>=10.0
PyPDF2>=3.0

# Google & External APIs
google-auth>=2.23
requests>=2.31

# Email
sib-api-v3-sdk>=3.0.0

# Production Server
gunicorn>=21.2

# Security & Performance
whitenoise>=6.11
